max77818-charger: fix dtls[] data type
The dtls[] data type should be u8, so that regmap_buck_read() can read
correct DTLS_00 ~ DTLS02 values into dtls[3].

+++ drivers/power/supply/max77818-charger.c
@@ -179,7 +179,8 @@ struct max77818_charger {
 	int chgin_irq;
 	int wcin_irq;
 	int chg_int;
-	int dtls[3];
+
+	u8 dtls[3];
 
 	int present;
 	int health;
